seperated functions out of the components
