+++ client/components/Navbar.jsx
@@ -2,15 +2,15 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { HashRouter as Router, Link } from 'react-router-dom'
 
-import { menuNavigation, menuVisable } from '../actions'
-
 import NavbarMenu from './NavbarMenu.jsx'
+
+import { openMenu } from './component_functions/openMenu'
 
 let Navbar = (props) => (
   <div className='header-menu-container'>
     <div className='header-container'>
       <div className='hambuger-container' onClick={() => openMenu(props)}>
-        { props.menuVisableState ? <i className='fa fa-times' aria-hidden='true' /> : <i className='fa fa-bars' aria-hidden='true' /> }
+        { props.menuVisable ? <i className='fa fa-times' aria-hidden='true' /> : <i className='fa fa-bars' aria-hidden='true' /> }
       </div>
       <Router>
         <div className='title-container'>
@@ -27,21 +27,15 @@ let Navbar = (props) => (
         </div>
       </Router>
     </div>
-    { props.menuState ? <NavbarMenu /> : '' }
+    { props.menuVisable ? <NavbarMenu /> : '' }
   </div>
 )
 
 function mapStateToProps (state) {
   return {
     dispatch: state.dispatch,
-    menuState: state.menuState,
-    menuVisableState: state.menuVisableState
+    menuVisable: state.menuVisableState
   }
-}
-
-function openMenu (props) {
-  props.dispatch(menuNavigation())
-  props.dispatch(menuVisable())
 }
 
 export default connect(mapStateToProps)(Navbar)
