+++ client/components/NavbarMenu.jsx
@@ -2,16 +2,16 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { HashRouter as Router, Link } from 'react-router-dom'
 
-import { menuNavigation, menuVisable, dashboardTab } from '../actions'
+import { goToPage } from './component_functions/goToPage'
 
 let NavbarMenu = (props) => (
   <div className='navigation-menu-container'>
     <Router>
       <ul>
-        <li className='navigation-item' onClick={() => openMenu(props)}><Link to='/login'>Register/Login</Link></li>
-        <li className='navigation-item' onClick={() => openMenu(props)}><Link to='/dashboard'>Your Dashboard</Link></li>
-        <li className='navigation-item' onClick={() => openMenu(props)}><Link to='/list-all'>Borrow an item</Link></li>
-        <li className='navigation-item' onClick={() => openMenu(props)}><Link to='/lender-form'>Lend an item</Link></li>
+        <li className='navigation-item' onClick={() => goToPage(props)}><Link to='/login'>Register/Login</Link></li>
+        <li className='navigation-item' onClick={() => goToPage(props)}><Link to='/dashboard'>Your Dashboard</Link></li>
+        <li className='navigation-item' onClick={() => goToPage(props)}><Link to='/list-all'>Borrow an item</Link></li>
+        <li className='navigation-item' onClick={() => goToPage(props)}><Link to='/lender-form'>Lend an item</Link></li>
       </ul>
     </Router>
   </div>
@@ -22,12 +22,6 @@ function mapStateToProps (state) {
     dispatch: state.dispatch,
     menuState: state.menuState
   }
-}
-
-function openMenu (props) {
-  props.dispatch(menuNavigation())
-  props.dispatch(menuVisable())
-  props.dispatch(dashboardTab('MySharingDetails'))
 }
 
 export default connect(mapStateToProps)(NavbarMenu)
