+++ client/components/component_functions/goToPage.js
@@ -1,0 +1,8 @@
+import { menuVisable, dashboardTab } from '../../actions'
+
+export { goToPage }
+
+function goToPage (props) {
+  props.dispatch(menuVisable())
+  props.dispatch(dashboardTab('MySharingDetails'))
+}
