+++ client/components/component_functions/openMenu.js
@@ -1,0 +1,7 @@
+import { menuVisable } from '../../actions'
+
+export { openMenu }
+
+function openMenu (props) {
+  props.dispatch(menuVisable())
+}
